+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Updated Theme to use Studio colors ([#838])
 * Improved patch visualizer UX ([#883])
 * Added experimental setting for Auto Connect in playtests ([#840])
+* Improved settings UI ([#886])
+* `Open Scripts Externally` option can now be changed while syncing ([#911])
 * Projects may now specify rules for syncing files as if they had a different file extension. ([#813], [#909])
  	This is specified via a new field on project files, `syncRules`:
 
@@ -68,9 +70,11 @@
 [#838]: https://github.com/rojo-rbx/rojo/pull/838
 [#840]: https://github.com/rojo-rbx/rojo/pull/840
 [#883]: https://github.com/rojo-rbx/rojo/pull/883
+[#886]: https://github.com/rojo-rbx/rojo/pull/886
 [#893]: https://github.com/rojo-rbx/rojo/pull/893
 [#903]: https://github.com/rojo-rbx/rojo/pull/903
 [#909]: https://github.com/rojo-rbx/rojo/pull/909
+[#911]: https://github.com/rojo-rbx/rojo/pull/911
 
 ## [7.4.1] - February 20, 2024
 * Made the `name` field optional on project files ([#870])